# packages/contracts/openapi/ceerion-mail.yml
openapi: 3.1.0
info:
  title: CEERION Mail API
  version: 1.0.0
  description: Backend API for CEERION Mail application with comprehensive authentication and mail handling
  contact:
    name: CEERION Team
  license:
    name: MIT
servers:
  - url: http://localhost:4000
    description: Development server
  - url: https://api.mail.ceerion.com
    description: Production server
security:
  - bearerJWT: []
paths:
...
  /auth/login:
    post:
      summary: User login with email/password
      operationId: login
      security: []
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              description: Refresh token cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimit'
...
  /auth/refresh:
    post:
      summary: Refresh access token
      operationId: refreshToken
      security:
        - cookieRefresh: []
      tags: [Authentication]
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 JWT access token
    cookieRefresh:
      type: apiKey
      in: cookie
      name: refresh_token
      description: Refresh token in HTTP-only cookie
    # Note: WebAuthn would be implemented via custom headers/body
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        emailVerified:
          type: boolean
        mfaEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, email, name, emailVerified, mfaEnabled, createdAt, updatedAt]
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        mfaCode:
          type: string
          pattern: '^[0-9]{6}$'
        rememberMe:
          type: boolean
          default: false
      required: [email, password]
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        expiresIn:
          type: integer
        user:
          $ref: '#/components/schemas/User'
        requiresMFA:
          type: boolean
        mfaToken:
          type: string
      required: [accessToken, expiresIn, user]
    RefreshResponse:
      type: object
      properties:
        accessToken:
          type: string
        expiresIn:
          type: integer
      required: [accessToken, expiresIn]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
        details:
          type: object
      required: [error, message]
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    RateLimit:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          description: Request limit per window
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Requests remaining in window
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Window reset time (Unix timestamp)
          schema:
            type: integer
        Retry-After:
          description: Seconds until retry allowed
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...